% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_data.R
\docType{data}
\name{env_data}
\alias{env_data}
\title{Environmental variable data provided by ocean monitoring in Ecuador}
\format{
\subsection{\code{env_data}}{

A data frame with 44 rows and 7 columns:
\describe{
\item{T}{Temperature, ºC}
\item{S}{Salinity, ups}
\item{OD}{Dissolved oxygen dissolved, mg.L-1}
\item{NO2}{Nitrite, ug-at.L-1}
\item{NO3}{Nitrate, ug-at.L-1}
\item{PO4}{phosphate, ug-at.L-1}
\item{SiO4}{Silicate, ug-at.L-1}
}
}
}
\source{
Oceanographic Institute of the Navy of Ecuador (INOCAR)
}
\usage{
env_data
}
\description{
Environmental data from physical-chemical and biological data sampling at four stations
located 10 miles from the coastal profile of continental Ecuador.
}
\keyword{datasets}
