% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiq_data.R
\docType{data}
\name{etiq_data}
\alias{etiq_data}
\title{Labeling of environmental data sets and phytoplankton species \strong{env_data}, \strong{spe_data} provided by ocean monitoring in Ecuador}
\format{
\subsection{\code{etiq_data}}{

A data frame with 44 rows and 3 columns:
\describe{
\item{ID}{Row identifier.}
\item{space}{Location from which the data were taken: 4 provinces of Ecuador.}
\item{month}{Month when measurement was made.}
}
}
}
\source{
Oceanographic Institute of the Navy of Ecuador (INOCAR).
}
\usage{
etiq_data
}
\description{
Location and time of measurement of physico-chemical and biological variables.
Data sorted according to \link{env_data} and \link{spe_data} rows.
}
\keyword{datasets}
