% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HJSTATICO.R
\name{HJSTATICO}
\alias{HJSTATICO}
\title{HJ-STATICO method}
\usage{
HJSTATICO(x, y, condition, bplot, preprocessing, compx, compy)
}
\arguments{
\item{x}{dataframe. A set of k matrices with (np) dimensions and juxtaposed in vertical order containing the data of the k conditions.}

\item{y}{dataframe. A set of k matrices with (nq) dimensions and juxtaposed in vertical order containing the data of the k conditions.}

\item{condition}{factor. The k condition descriptions. The third path for the analysis.}

\item{bplot}{logical. TRUE by default, it will estimate the components of the Biplot for the compromise and for the intra-structure considering the HJ-Biplot perspective from Galindo, 1986.}

\item{preprocessing}{Standardization method to be used. "B.Total" performs total standardization and "B.Partial" partial standardization of Bouroche (1975) (library ade4) and "A.Norma" standardization of Abdi et al. (2012).}

\item{compx}{Indicates the component of x to be used for the representation of the compromise and intra-structure plots on the bi-factor plane.}

\item{compy}{Indicates the component of y to be used for the representation of the compromise and intra-structure plots on the bi-factor plane.}
}
\value{
A list containing the following components:
\itemize{
\item \emph{X:}	The standardized matrix of x, resulting from using the Prep.Tables() function from the 'staticoex4' library, with the preprocessing argument indicating the type of standardization to be applied.
\item \emph{Y:}	The \strong{centered} matrix of y (column-wise), but not standardized.
\item \emph{CoIATables:}	A matrix with the k-tables of coinertia arranged vertically.
\item \emph{Compromise:}	The matrix resulting from linear combination weighting the first’s eigenvector components and k-tables of coinertia (The stable part).
\item \emph{HJB:}	The representation quality metrics and contributions for the compromise matrix obtained using the HJ-Biplot method.
\item \emph{HJBK:}	The representation quality metrics and contributions for the intra-structure obtained using the HJ-Biplot method (the projection of each Coinertia table onto the compromise).
\item \emph{Rv:}	The vectorial correlations matrix.
\item \emph{Interstructure:}	The eigenvalues information about the inertia achieved in the inter-structure.
\item \emph{CompromiseSVD:}	The eigenvalues information from the stable part of the co-structure of k-tables.
\item \emph{K_Indexs:}	K-tables information related with the compromise matrix. Weights and square cosines of each k-table.
}
}
\description{
Analysis of the stable part of the co-structure between two three-way data sets X(npk) and Y(nqk), getting the relationship quantification and biplots. Multivariate method that combines STATICO methods for multi-way data analysis and HJ-Biplot.
}
\details{
The dimensions of the commitment are determined by the rank of the matrix.
}
\references{
http://hdl.handle.net/10366/149381
}
\author{
Mariela Gonzalez Narvaez, Ph.D., María José Fernández Gomez, Susana Mendes, Purificación Galindo-Villardón, Ph.D., Gema Zambrano Zambrano, Ysaí Ronquillo Mora, George Acosta Chong
}
