% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepTables.R
\name{Prep.Tables}
\alias{Prep.Tables}
\title{Table preprocessing: Centering and Normalizing}
\usage{
Prep.Tables(x, y, condition, preprocessing)
}
\arguments{
\item{x}{A dataframe with k matrices juxtaposed in vertical order containing the data of the P variables (environmental information).}

\item{y}{A dataframe with k matrices juxtaposed in vertical order containing the data of the Q variables (species information).}

\item{condition}{Factor, the k condition descriptions that will define the third path for the analysis.}

\item{preprocessing}{A string indicating the standardization method to be used: "B.Total" performs total standardization and "B.Partial" partial standardization of Bouroche (1975) (library ade4) and "A.Norma" standardization of Abdi et al. (2012). See details.}
}
\value{
Returns a list-type object with three elements:
\itemize{
\item \emph{X:} It is the normalized matrix corresponding to “x” (environmental matrix).
\item \emph{Y:} It is the \strong{centered} matrix corresponding to “y” (species matrix).
\item \emph{Index:} A data.frame with the name of k-tables, the number of observations and the index of start and finish for each table.
}
}
\description{
Scaling of two dataframes: environmental and species variables. There are three options for preprocessing of tables regarding centering and normalizing.
}
\details{
Table preprocessing function that considers a third way or categorical variable that divides data into k tables. That is, the k pairs of individual tables must be appended both in the matrix of the environmental variables and in the species data and, in both cases, have the same number of observations, since an index table is built and runs through the same positions of the observations for both matrices. There are three options for matrix scaling: centering and normalizing to remove the effects of different means and scales.
•	“B.Partial” center and normalize the columns for each k section, ensuring that the initial k matrices have zero mean and unit variance.
•	“B.Total” center the columns for each k section, then normalize for the total column, ensuring that the initial k-matrices have a zero mean, and the variance of the total column is one.
•	“A.Norma” This method operates similarly to dividing each element of the total matrix by the square root of the total sum of squared elements, ensuring that the total variance of each k-matrix is equal to one and that all K matrices participate equally in the analysis.
It is worth mentioning that X is normalized, and Y is only centered, due to X is taken as variables or measurements with possible different scales and Y as frequency data (species frequencies).
}
\examples{
Prep.Tables(x=env_data, y=spe_data, condition=etiq_data$space, preprocessing = "A.Norma")

}
\references{
http://hdl.handle.net/10366/149381
}
\seealso{
\code{\link[=HJSTATICO]{HJSTATICO()}}, ade4::statico, MixSTATICO()
}
\author{
Mariela Gonzalez Narvaez, Ph.D., María José Fernández Gomez, Susana Mendes, Purificación Galindo-Villardón, Ph.D., Gema Zambrano Zambrano, Ysaí Ronquillo Mora, George Acosta Chong
}
