% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_data.R
\docType{data}
\name{spe_data}
\alias{spe_data}
\title{Phytoplankton species data provided by ocean monitoring in Ecuador}
\format{
\subsection{\code{spe_data}}{

A data frame with 44 rows and 23 columns:
\describe{
\item{e1}{affinis (\emph{Chaetoceros})}
\item{e2}{curvisetus (\emph{Chaetoceros})}
\item{e3}{didymus (\emph{Chaetoceros})}
\item{e4}{sinensis (\emph{Hemiaulus})}
\item{e5}{danicus (\emph{Leptocylindrus})}
\item{e6}{brightwellii (\emph{Ditylum})}
\item{e7}{fragilissimus (\emph{Dactyliosolen})}
\item{e8}{striata (\emph{Guinardia})}
\item{e9}{imbricata (\emph{Rhizosolenia})}
\item{e10}{alata (\emph{Proboscia})}
\item{e11}{antarcticus (\emph{Dactyliosolen})}
\item{e12}{costatum (\emph{Skeletonema})}
\item{e13}{borealis (\emph{Lauderia})}
\item{e14}{sp. (\emph{Thalassiosira})}
\item{e15}{sp. (\emph{Nitzschia})}
\item{e16}{longissima (\emph{Nitzschia})}
\item{e17}{pungens (\emph{Pseudo-nitzschia})}
\item{e18}{sp. (\emph{Navicula})}
\item{e19}{membranacea (\emph{Stauroneis})}
\item{e20}{nitzschioides (\emph{Thalassionema})}
\item{e21}{sp. (\emph{Gymnodinium})}
\item{e22}{sp. (\emph{Gyrodinium})}
\item{e23}{rubrum (\emph{Mesodinium})}
}
}
}
\source{
Oceanographic Institute of the Navy of Ecuador (INOCAR)
}
\usage{
spe_data
}
\description{
Measurement of the abundance \if{html}{\out{<cells L-1>}} of 23 phytoplankton species
from sampling data at four stations located 10 miles from
the coastal profile of the continental equator.
Data sorted according to \link{env_data} and \link{etiq_data} rows.
}
\keyword{datasets}
